# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TODO-LIST API",
    "description": "開發具備CRUD功能的TODO-LIST API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
  "tags": [
    {
      "name": "Todos",
      "description": "待辦事項",
      "externalDocs": {
        "description": "",
        "url": ""
      }
    }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "tags": ["Todos"],
        "summary": "取得所有待辦事項",
        "responses": {
          "200": {
            "description": "成功取得待辦事項列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Todos"],
        "summary": "新增待辦事項",
        "requestBody": {
          "description": "新增待辦事項的內容",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Todo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功新增待辦事項",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "patch": {
        "tags": ["Todos"],
        "summary": "更新待辦事項完成狀態",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "待辦事項完成狀態",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "completed": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功更新待辦事項完成狀態"
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "刪除指定待辦事項",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功刪除指定待辦事項"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Todo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "task": {
            "type": "string",
            "example": "舉辦一場空氣吉他比賽，看看誰能演奏得最好。"
          },
          "completed": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  }
}